feat: add consensus engine integration
- Created WaveFPCEngine for FPC fast-path consensus
- Added VerkleVMIntegration bridging VMs with Verkle+FPC
- Integrated with node/snow context
- Full Verkle witness validation support
- Tests passing with ~100ns fast path performance

+++ network/throttling/inbound_resource_throttler.go
@@ -140,12 +140,12 @@ func (t *systemThrottler) Acquire(ctx context.Context, nodeID ids.NodeID) {
 		target := t.targeter.TargetUsage()
 		// Get actual usage for this node.
 		usage := t.tracker.Usage(nodeID, now)
-		if usage <= target {
+		if usage <= float64(target) {
 			return
 		}
 		// See how long it will take for actual usage to drop to the target,
 		// assuming this node uses no more resources.
-		waitDuration := t.tracker.TimeUntilUsage(nodeID, now, target)
+		waitDuration := t.tracker.TimeUntilUsage(nodeID, now, float64(target))
 		if waitDuration < epsilon {
 			// If the amount of time until we reach the target is very small,
 			// just return to avoid a situation where we excessively re-check.
